requirements.txt
# Core ML/DL - Native Windows GPU Support
tensorflow-gpu==2.10.1
tensorflow-estimator==2.10.0
tensorboard==2.10.0
tensorflow-addons==0.19.0
tensorflow-datasets==4.9.2
tensorflow-io-gcs-filesystem==0.28.0

# Computer Vision
opencv-python-headless==4.7.0.72
Pillow>=9.0.0,<10.0.0

# Data Science
numpy>=1.21.0,<1.24.0
pandas>=1.3.0,<2.0.0
scikit-learn==1.2.2
scipy>=1.7.0,<1.11.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
altair>=4.0.0

# Web App
streamlit==1.50.0
streamlit-webrtc==0.63.5

# Utilities
pyyaml>=6.0
requests>=2.27.0
h5py>=3.1.0,<3.9.0
joblib>=1.1.0
tqdm>=4.62.0
attrs>=21.0.0
jinja2>=3.0.0
werkzeug>=2.0.0,<3.0.0